Move authorization package into priorauth to fix tests
