+++ src/main/java/org/hl7/davinci/priorauth/authorization/AuthEndpoint.java
@@ -1,4 +1,4 @@
-package org.hl7.davinci.authorization;
+package org.hl7.davinci.priorauth.authorization;
 
 import java.math.BigInteger;
 import java.security.KeyFactory;
