+++ src/main/java/org/hl7/davinci/priorauth/authorization/WellKnownEndpoint.java
@@ -1,4 +1,4 @@
-package org.hl7.davinci.authorization;
+package org.hl7.davinci.priorauth.authorization;
 
 import java.util.HashMap;
 import java.util.Map;
